requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-tcs34725==3.3.20
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.73.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==24.2.0
awscli==1.33.44
backoff==2.2.1
binho-host-adapter==0.1.6
boto3==1.34.123
botocore==1.34.162
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
fastapi==0.110.3
filelock==3.15.4
flatbuffers==20181003210633
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==3.8
imageio==2.35.1
inference==0.16.3
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.15.1
openai==1.42.0
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.4.0
pluggy==1.5.0
prettytable==3.11.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.0.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
PuLP==2.9.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydot==3.0.1
pyftdi==0.55.4
Pygments==2.18.0
pyparsing==3.1.3
pyserial==3.5
pytest==8.3.2
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.20.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
skypilot==0.5.0
sniffio==1.3.1
starlette==0.37.2
structlog==24.4.0
supervision==0.22.0
sympy==1.13.2
sysv_ipc==1.1.0
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.8.10
time-machine==2.15.0
tldextract==5.1.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.82
ultralytics-thop==2.0.5
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.9.4
zxing-cpp==2.2.0
